chore: split up validator file for clarity
