+++ packages/form-core/src/internals.ts
@@ -7,7 +7,7 @@ export * from './FieldGroup/FieldGroupApi.lib'
 export * from './utils.lib'
 export * from './types.lib'
 export * from './FieldApi/RootFieldApi.lib'
-export * from './validation.lib'
+export * from './validation'
 export * from './listeners.lib'
 export * from './FieldApi/linked-fields.lib'
 export * from './standardSchema.lib'
