+++ packages/form-core/src/ssr.lib.ts
@@ -3,7 +3,7 @@ import { defaultInternalBaseFieldMeta } from './FieldApi/fieldState.lib'
 import { visitAllFormFields } from './FieldApi/fieldTraversal.lib'
 import { parseStandardSchemaIssues } from './standardSchema.lib'
 import { cancelPipelineCache, createPipelineCache, evaluate } from './utils.lib'
-import { runValidatorPipeline } from './validation.lib'
+import { runValidatorPipeline } from './validation'
 import { createErrorMap } from './validation.public'
 import { devtools } from './devtoolsBridge.lib'
 import type { FormOptions } from './FormApi/FormApi.public'
